// src/types/broker/BrokerType.d.ts
export interface BrokerType {
	business_address_line_1: string;
	business_address_line_2: string;
	business_city: string;
	business_country: string;
	business_kifle_ketema: string;
	business_license_number: string;
	business_state: string;
	business_type: string;
	business_zip_code: string;
	company_name: string;
	date_of_birth: string;
	email_address: string;
	first_name: string;
	gender: string;
	last_name: string;
	middle_initial: string;
	phone_number: string;
	tax_identification_number: string;
	broker_type: string;
	license_state: string;
	license_issued_date: string;
	license_expired_date: string;
	in_network: boolean;
	monthly_premium: number;
	yearly_premium: number;
}